fix(auth): post-review hardening — 2FA ordering on standard path, extended reserved slugs, int casts, 2FA intrusion guard
- Move maybeGenerateTwoFactorCode() after $assos->count() === 0 guard on
  the standard /login path: prevents leaking credential validity by sending
  a 2FA email to a user who has no active association membership.
- Add test: POST /login with 2FA-enabled user and no asso → no email sent,
  redirect login with 'Compte non rattaché à une association.' error.
- Add 7 missing reserved slugs to config/tenancy.php (password, verify-email,
  confirm-password, forgot-password, reset-password, two-factor, up).
  ReservedSlugTest dataProvider auto-covers them (no test changes needed).
- Apply (int) casts on session('current_association_id') comparisons in
  LoginBrandedValidationTest (scenarios 1 and 4) per project convention.
- Fix SlugLoginIntrusionTest scénario 4b: marie now has two_factor_method
  TwoFactorMethod::Email + two_factor_confirmed_at so the guard-before-2FA
  invariant is actually tested with 2FA genuinely enabled.

+++ tests/Feature/Auth/LoginBrandedValidationTest.php
@@ -52,7 +52,7 @@ test('scenario 1: user belonging to SVS can login via /svs/login and lands on da
 
     $this->assertAuthenticated();
     $response->assertRedirect(route('dashboard'));
-    expect(session('current_association_id'))->toBe($this->svs->id);
+    expect((int) session('current_association_id'))->toBe((int) $this->svs->id);
 });
 
 test('scenario 2: user not belonging to SVS is rejected on /svs/login with French error, session not authenticated', function () {
@@ -89,7 +89,7 @@ test('scenario 4: multi-asso user logging in via /svs/login gets SVS context eve
 
     $this->assertAuthenticated();
     $response->assertRedirect(route('dashboard'));
-    expect(session('current_association_id'))->toBe($this->svs->id);
+    expect((int) session('current_association_id'))->toBe((int) $this->svs->id);
     // derniere_association_id must be updated to SVS as well
     $this->multi->refresh();
     expect($this->multi->derniere_association_id)->toBe($this->svs->id);
